+++ mycodo/mycodo_flask/admin_routes.py
@@ -11,6 +11,7 @@ from flask import (Blueprint,
                    request,
                    session,
                    url_for)
+from flask_babel import gettext
 from pkg_resources import parse_version
 
 from utils.statistics import return_stat_file_dict
@@ -18,6 +19,7 @@ from utils.system_pi import internet
 from utils.github_release_info import github_releases
 
 from mycodo import flaskforms
+from mycodo import flaskutils
 from mycodo.mycodo_flask.general_routes import (before_blueprint_request,
                                                 inject_mycodo_version,
                                                 logged_in)
@@ -44,7 +46,7 @@ def admin_backup():
         return redirect(url_for('general_routes.home'))
 
     if session['user_group'] == 'guest':
-        flash("Guests are not permitted to view backups.", "error")
+        flaskutils.deny_guest_user()
         return redirect(url_for('general_routes.home'))
 
     formBackup = flaskforms.Backup()
@@ -83,7 +85,7 @@ def admin_statistics():
         return redirect(url_for('general_routes.home'))
 
     if session['user_group'] == 'guest':
-        flash("Guests are not permitted to view statistics.", "error")
+        flaskutils.deny_guest_user()
         return redirect(url_for('general_routes.home'))
 
     try:
@@ -101,13 +103,12 @@ def admin_upgrade():
         return redirect(url_for('general_routes.home'))
 
     if session['user_group'] == 'guest':
-        flash("Guests are not permitted to view the upgrade panel.",
-              "error")
+        flaskutils.deny_guest_user()
         return redirect(url_for('general_routes.home'))
 
     if not internet():
-        flash("Upgrade functionality is disabled because an internet "
-              "connection was unable to be detected.", "error")
+        flash(gettext("Upgrade functionality is disabled because an internet "
+                      "connection was unable to be detected"), "error")
         return render_template('admin/upgrade.html',
                                is_internet=False)
 
@@ -125,11 +126,12 @@ def admin_upgrade():
 
     if upgrade:
         if upgrade == 1:
-            flash("An upgrade is currently in progress. Please wait for it to"
-                  " finish.", "error")
+            flash(gettext("An upgrade is currently in progress. Please wait "
+                          "for it to finish"), "error")
         elif upgrade == 2:
-            flash("There was an error encountered during the upgrade process."
-                  " Check the upgrade log for details.", "error")
+            flash(gettext("There was an error encountered during the upgrade "
+                          "process. Check the upgrade log for details."),
+                  "error")
         return render_template('admin/upgrade.html',
                                upgrade=upgrade)
 
@@ -164,10 +166,10 @@ def admin_upgrade():
                                 path=INSTALL_DIRECTORY),
                              shell=True)
             upgrade = 1
-            flash("The upgrade has started. The daemon will be "
-                  "stopped during the upgrade.", "success")
+            flash(gettext("The upgrade has started. The daemon will be "
+                          "stopped during the upgrade."), "success")
         else:
-            flash("You cannot upgrade if an upgrade is not available",
+            flash(gettext("You cannot upgrade if an upgrade is not available"),
                   "error")
 
     return render_template('admin/upgrade.html',